montage/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Admin==1.6.1
Flask-Mail==0.9.1
WTForms==3.1.0
SQLAlchemy==2.0.23
bcrypt==4.1.2
email-validator==2.1.0
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
